+++ plugins/decl_hdf5/CHANGELOG.md
@@ -7,6 +7,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 ## [Unreleased]
 
 ### Added
+* Support specifying a policy when the file or dataset to write to already
+  exists [#121](https://gitlab.maisondelasimulation.fr/pdidev/pdi/-/issues/121)
+
 
 ### Changed